Begin decoupling tests from main algorthim. Still buggy.
