+++ include/cylaks/linear_spring.hpp
@@ -10,7 +10,7 @@ class BindingSite;
 // LinearSpring: Hookean; works to hold two points at some rest distance
 // Note: currently contains some junk related to angular springs; to fix
 class LinearSpring : public Object {
-private:
+protected:
   double k_slack_{0.0};  // For when r < r_0; pN/nm
   double k_spring_{0.0}; // For when r > r_0; pN/nm
   double k_rot_{0.0};
@@ -30,7 +30,7 @@ public:
   double theta_rest_{0.0};
   double theta_max_{0.0};
 
-private:
+protected:
   void SetCutoffs() {
     // Find cutoff values by setting a max Boltzmann weight of 1e3
     double max_weight{1e3};
