+++ src/filament_manager.cpp
@@ -3,8 +3,6 @@
 
 void FilamentManager::SetParameters() {
 
-  using namespace Params;
-  using namespace Filaments;
   // Filaments are mobile so long as at least 1 dimension is enabled
   for (int i_dim{0}; i_dim < _n_dims_max; i_dim++) {
     if (Params::Filaments::translation_enabled[i_dim]) {
@@ -17,65 +15,46 @@ void FilamentManager::SetParameters() {
   threshold_ = std::pow(2, 1.0 / 6.0) * sigma_;
   n_bd_iterations_ = Params::Filaments::n_bd_per_kmc;
   dt_eff_ = Params::dt / n_bd_iterations_;
-  weight_neighbs_bind_.resize(_n_neighbs_max + 1);
-  weight_neighbs_unbind_.resize(_n_neighbs_max + 1);
-  for (int n_neighbs{0}; n_neighbs <= _n_neighbs_max; n_neighbs++) {
-    double dE{n_neighbs * -Xlinks::neighb_neighb_energy};
-    weight_neighbs_bind_[n_neighbs] = exp(-(1.0 - _lambda_neighb) * dE);
-    weight_neighbs_unbind_[n_neighbs] = exp(_lambda_neighb * dE);
-  }
 }
 
 void FilamentManager::GenerateFilaments() {
 
-  using namespace Sys;
-  using namespace Params;
-  proto_.resize(Filaments::count);
-  for (int i_fil{0}; i_fil < proto_.size(); i_fil++) {
-    proto_[i_fil].Initialize(_id_site, n_unique_objects_++, i_fil);
+  // using namespace Params;
+  protofilaments_.resize(Params::Filaments::count);
+  for (int i_fil{0}; i_fil < protofilaments_.size(); i_fil++) {
+    protofilaments_[i_fil].Initialize(_id_site, Sys::n_objects_++, i_fil);
   }
-  for (auto &&pf : proto_) {
+  for (auto &&pf : protofilaments_) {
     for (auto &&site : pf.sites_) {
       sites_.emplace_back(&site);
     }
   }
-  if (proto_.size() == 2) {
-    proto_[0].neighbor_ = &proto_[1];
-    proto_[1].neighbor_ = &proto_[0];
+  if (protofilaments_.size() == 2) {
+    protofilaments_[0].neighbor_ = &protofilaments_[1];
+    protofilaments_[1].neighbor_ = &protofilaments_[0];
   }
   // FIXME gamma_rot is invalid for MTs that are too short; need better
   // expression --- do not include for now
+  using namespace Sys;
   int n_dims{2}; // 3};
   Vec<double> D(n_dims, 0.0);
   Log("  Filament variables calculated post-initialization:\n");
-  for (auto const &pf : proto_) {
+  for (auto const &pf : protofilaments_) {
     Log("   length[%i] = %g nm\n", pf.index_, pf.length_);
   }
-  for (auto const &pf : proto_) {
-    Log("    D_par[%i] = %g nm^2/s\n", pf.index_, kbT / pf.gamma_[0]);
+  for (auto const &pf : protofilaments_) {
+    Log("    D_par[%i] = %g nm^2/s\n", pf.index_, Params::kbT / pf.gamma_[0]);
   }
-  for (auto const &pf : proto_) {
-    Log("     D_perp[%i] = %g nm^2/s\n", pf.index_, kbT / pf.gamma_[1]);
+  for (auto const &pf : protofilaments_) {
+    Log("     D_perp[%i] = %g nm^2/s\n", pf.index_, Params::kbT / pf.gamma_[1]);
   }
-  for (auto const &pf : proto_) {
+  for (auto const &pf : protofilaments_) {
     Log("      gamma_par[%i] = %g nm^2/s\n", pf.index_, pf.gamma_[0]);
   }
-  for (auto const &pf : proto_) {
+  for (auto const &pf : protofilaments_) {
     Log("       gamma_perp[%i] = %g nm^2/s\n", pf.index_, pf.gamma_[1]);
   }
   Log("\n");
-  /*
-  for (int i_fil{0}; i_fil < proto; i_fil++) {
-    Sys::Log("mt #%i\n", proto_[i_fil].index_);
-    Sys::Log("%zu & %zu\n", proto_[i_fil].plus_end_->pos_.size(),
-             proto_[i_fil].minus_end_->pos_.size());
-    Sys::Log("plus-end: (%g, %g)\n", proto_[i_fil].plus_end_->pos_[0],
-             proto_[i_fil].plus_end_->pos_[1]);
-    Sys::Log("minus_end: (%g, %g)\n", proto_[i_fil].minus_end_->pos_[0],
-             proto_[i_fil].minus_end_->pos_[1]);
-  }
-  exit(1);
-  */
 }
 
 bool FilamentManager::AllFilamentsImmobile() {
@@ -83,7 +62,7 @@ bool FilamentManager::AllFilamentsImmobile() {
   if (!mobile_) {
     return true;
   }
-  for (auto const &pf : proto_) {
+  for (auto const &pf : protofilaments_) {
     if (Sys::i_step_ >= pf.immobile_until_) {
       return false;
     }
@@ -92,13 +71,14 @@ bool FilamentManager::AllFilamentsImmobile() {
 }
 
 void FilamentManager::UpdateForces() {
-  for (auto &&pf : proto_) {
+  for (auto &&pf : protofilaments_) {
     for (int i_dim{0}; i_dim < _n_dims_max; i_dim++) {
       pf.force_[i_dim] = Params::Filaments::f_applied[i_dim];
     }
     pf.torque_ = 0.0;
   }
   // Temporary hacky 1-D wca potential implementation
+  // !FIXME: add flag to yaml file to enable/disable this
   /*
   if (proto_.size() == 2 and proteins_->xlinks_.active_) {
     double r{proto_[1].pos_[1] - proto_[0].pos_[1]};
@@ -113,7 +93,9 @@ void FilamentManager::UpdateForces() {
     }
   }
   */
-  proteins_->UpdateExtensions();
+  // proteins_->UpdateExtensions();
 }
 
-void FilamentManager::UpdateLattice() { proteins_->UpdateLatticeDeformation(); }
+void FilamentManager::UpdateLattice() {
+  // proteins_->UpdateLatticeDeformation();
+}
